CheckInputs

+++ Entities/Entity.java
@@ -83,7 +83,7 @@ public abstract class Entity{
     * Igualar {@code (hitbox.x, hitbox.y)} a las {@code (x,y)} "ancladas"
      */
     public void resetHitboxPos(){
-    this.hitbox.x = (int)this.x;
-    this.hitbox.y = (int)this.y;
+    this.hitbox.x = this.x;
+    this.hitbox.y = this.y;
    }
 }
